+++ src/Theme.h
@@ -32,6 +32,7 @@ COLORREF ThemeNotificationsProgressColor();
 bool ThemeColorizeControls();
 bool ThemeUsesDarkChrome();
 bool ThemeUsesBlackChrome();
+bool ThemeUsesOriginalPageColors();
 bool IsCurrentThemeDefault();
 COLORREF AccentColor(COLORREF col, int light, int dark = 0);
 void FreeThemes();